requirements.txt
amqp==2.1.4
Babel==2.3.4
billiard==3.5.0.2
celery==4.0.2
Django==1.10.4
django-cors-headers==1.3.1
djangorestframework==3.5.3
flower==0.9.1
kombu==4.0.2
oauthlib==2.0.1
psycopg2==2.6.2
pytz==2016.10
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
tornado==4.2
tweepy==3.5.0
uWSGI==2.0.14
vine==1.1.3
